Add basic documentation and split out classes.
